move helm/cilium to dedicated file
